requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
sniffio==1.3.0
surrealdb @ git+https://github.com/Scurrra/surrealdb.py.git@68853a77917562be20375843dc0dcfd8bad728e7
typing_extensions==4.7.1
websockets==10.4
yarl==1.9.2
